test(speech-typist): cover the two outward failures that are not Lemonade's
clippy noticed the fake host's failure switches were never flipped, which was
the honest signal that two paths in the core had no test: a capture device that
will not open (another application holding it exclusively, a revoked microphone
permission) and an injection the window refuses. Both are reported and leave the
tray broken; neither is silent, which is what matters — a person talking to a
key that stopped working has no other way to find out.

Both targets are now clippy-clean.

+++ speech-typist/src/win/mod.rs
@@ -288,7 +288,7 @@ extern "system" fn window_proc(hwnd: HWND, message: u32, wparam: WPARAM, lparam:
             LRESULT(0)
         }
         WM_COMMAND => {
-            on_command((wparam.0 & 0xFFFF) as usize);
+            on_command(wparam.0 & 0xFFFF);
             LRESULT(0)
         }
         WM_DESTROY => {
